# nextjs/public/openapi.yaml
openapi: 3.1.0
info:
  title: VidTempla API
  version: "1.0"
  description: |
    VidTempla is an API-first platform for managing YouTube video descriptions with templates, containers, and variables.
    The REST API lets AI agents securely manage YouTube channels without handling OAuth directly.

    ## Authentication
    All requests require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer vtk_your_key_here
    ```
    API keys have one of two permission levels:
    - **read** — can only fetch data (GET requests and POST /analytics/query)
    - **read-write** — can also create, update, and delete resources

    Generate keys from the dashboard at Settings > API Keys.

    ## Response Envelope
    Every response uses a standard envelope:
    ```json
    { "data": { ... }, "error": null, "meta": { ... } }
    ```

    ## Pagination
    All list endpoints use cursor-based pagination. Default limit is 50, max is 100.
    ```
    GET /videos?cursor=...&limit=50
    ```
  contact:
    name: VidTempla
    url: https://vidtempla.com
  license:
    name: Proprietary
servers:
  - url: https://vidtempla.com/api/v1
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key starting with vtk_. Generate from Settings > API Keys."
  schemas:
    ErrorResponse:
      type: object
      properties:
        data:
          type: "null"
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            suggestion:
              type: string
              description: Actionable suggestion for agents to self-correct
            status:
              type: integer
              description: HTTP status code
            meta:
              type: object
              additionalProperties: true
              description: Optional machine-readable context about the error
          required: [code, message, status]
      required: [data, error]
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: INVALID_API_KEY
              message: API key not found or has been revoked
              suggestion: Generate a new key from Settings > API Keys
              status: 401
    InsufficientPermissions:
      description: API key lacks write permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: INSUFFICIENT_PERMISSIONS
              message: This API key has read-only access
              suggestion: Create a new API key with read-write permission from Settings > API Keys
              status: 403
    QuotaExceeded:
      description: YouTube API quota or rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: QUOTA_EXCEEDED
              message: YouTube API quota exhausted
              suggestion: Retry with backoff. If daily quota is exhausted, wait for the Pacific midnight reset or contact support.
              status: 429
              meta:
                upstreamStatus: 403
                reasons: [quotaExceeded]
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred
              suggestion: Try again later
              status: 500
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /youtube/comments/{id}:
    get:
      operationId: listComments
      summary: List video comments
      description: Lists comment threads for a YouTube video. The `{id}` is a YouTube video ID.
      tags: [YouTube Comments]
      x-quota-cost: 1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: YouTube video ID (for listing comments) or YouTube comment ID (for deletion)
        - name: channelId
          in: query
          required: true
          schema:
            type: string
          description: YouTube channel ID (for auth)
        - name: maxResults
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Max results per page (max 100)
        - name: order
          in: query
          schema:
            type: string
            enum: [relevance, time]
            default: relevance
          description: Sort order (relevance for top comments, time for newest)
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination token from the previous page's meta.cursor
        - name: pageToken
          in: query
          schema:
            type: string
          description: Alias of cursor, kept for backwards compatibility
      responses:
        "200":
          description: List of comment threads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      description: YouTube commentThread resource
                  error:
                    type: "null"
                  meta:
                    type: object
                    properties:
                      cursor:
                        type: string
                        nullable: true
                        description: Pass as ?cursor= for the next page. Null on the last page.
                      hasMore:
                        type: boolean
                      total:
                        type: integer
                        nullable: true
                        description: Always null. YouTube reports no total for a comment listing.
                      quotaUnits:
                        type: integer
                        example: 1
        "400":
          description: Missing channelId or invalid order parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/QuotaExceeded"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteComment
      summary: Delete a comment
      description: |
        Deletes a YouTube comment permanently. The `{id}` is a YouTube comment ID.

        The comment's text is snapshotted to the edit log before the delete, because
        YouTube keeps no comment history and a deleted comment cannot be re-found —
        that row is the only surviving copy. Read it back with `GET /youtube/comments/edits`.
        For a comment the acting channel did not write, YouTube returns only the
        HTML-marked-up `textDisplay`, so the snapshot is an audit record, not a
        restorable copy.
      tags: [YouTube Comments]
      x-quota-cost: 51
      x-requires-write: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: YouTube comment ID
        - name: channelId
          in: query
          required: true
          schema:
            type: string
          description: YouTube channel ID (UC...) whose token signs the delete
        - name: videoId
          in: query
          schema:
            type: string
          description: Video the comment sits on, recorded on the edit-log row when known
      responses:
        "200":
          description: Comment deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
                        example: true
                      editId:
                        type: string
                        format: uuid
                        description: Edit-log row holding the comment's text as it was before the delete
                  error:
                    type: "null"
                  meta:
                    type: object
                    properties:
                      quotaUnits:
                        type: integer
                        example: 51
        "400":
          description: Missing channelId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InsufficientPermissions"
        "404":
          description: The comment or the channel was not found. Nothing was written to YouTube.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/QuotaExceeded"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /youtube/comments/bulk-update:
    post:
      operationId: bulkUpdateComments
      summary: Rewrite up to 40 comments
      description: |
        Rewrites the text of up to 40 comments authored by one channel in a single call —
        the way to replace a link everywhere it appears. Find the comments first with
        `GET /youtube/comments?searchTerms=...`, then send the IDs you chose.

        Comments are edited in place. Delete-and-repost would unpin them and discard
        their likes and original timestamps, so it is never done.

        The batch runs in three phases. It first reconciles any rows left `pending` by an
        interrupted earlier batch for the same channel; then it snapshots and validates
        every item; then it writes. If any comment is missing or was not authored by
        `channelId`, the batch aborts and **nothing** is written to YouTube.

        If the YouTube daily quota runs out mid-batch, earlier items stay applied, the
        failing item returns `QUOTA_EXCEEDED`, and the rest come back as `skipped` and were
        never attempted. Send the skipped IDs again after the reset to resume.

        Split larger sweeps into batches of 40 and loop.
      tags: [YouTube Comments]
      x-quota-cost: 51
      x-requires-write: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channelId, items]
              properties:
                channelId:
                  type: string
                  description: YouTube channel ID (UC...) that authored every comment in the batch
                items:
                  type: array
                  minItems: 1
                  maxItems: 40
                  items:
                    type: object
                    required: [id, text]
                    properties:
                      id:
                        type: string
                        description: YouTube comment ID
                      videoId:
                        type: string
                        description: Video the comment sits on, recorded on the edit-log row when known
                      text:
                        type: string
                        description: Replacement text for this comment
            example:
              channelId: "UCxxxxxxxxxxxxxxxxxxxxxx"
              items:
                - id: "UgxAbCdEfGhIjKlMnOp"
                  videoId: "dQw4w9WgXcQ"
                  text: "Course link updated: https://example.com/course"
      responses:
        "200":
          description: Batch finished. Inspect every item's status — some can have failed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        description: One entry per submitted item, in the order sent
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            status:
                              type: string
                              enum: [ok, error, skipped]
                              description: skipped means the item was never attempted, after a quota halt
                            error:
                              type: object
                              properties:
                                code:
                                  type: string
                                message:
                                  type: string
                      reconciled:
                        type: object
                        description: Rows left pending by an interrupted earlier batch that this call resolved
                        properties:
                          scanned:
                            type: integer
                          applied:
                            type: integer
                          failed:
                            type: integer
                          unknown:
                            type: integer
                      resetsAt:
                        type: string
                        format: date-time
                        description: Present only when a YouTube quota halt stopped the batch
                  error:
                    type: "null"
                  meta:
                    type: object
                    properties:
                      quotaUnits:
                        type: integer
                        description: 51 per item attempted, plus 1 per row reconciled
                        example: 102
        "400":
          description: Malformed body, more than 40 items, duplicate comment IDs, or a channelId that is not a UC... ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: |
            The API key is read-only, or a comment was not authored by `channelId`
            (`AUTHORSHIP_MISMATCH`). On an authorship mismatch the batch aborted and
            nothing was written to YouTube — drop that ID and send the batch again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: A comment in the batch no longer exists (`SNAPSHOT_FAILED`). Nothing was written to YouTube.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/QuotaExceeded"
        "500":
          $ref: "#/components/responses/InternalError"
...